+++ account-service/src/main/java/backend/microservices/service/CurrencyService.java
@@ -3,10 +3,10 @@ package backend.microservices.service;
 import java.math.BigDecimal;
 
 public interface CurrencyService {
-    BigDecimal mapTengeToRub(BigDecimal tenge);
-    BigDecimal mapTengeToUsd(BigDecimal tenge);
-    BigDecimal mapRubToTenge(BigDecimal rub);
-    BigDecimal mapRubToUsd(BigDecimal rub);
-    BigDecimal mapUsdToTenge(BigDecimal usd);
-    BigDecimal mapUsdToRub(BigDecimal usd);
+    Long mapTengeToRub(Long tenge);
+    Long mapTengeToUsd(Long tenge);
+    Long mapRubToTenge(Long rub);
+    Long mapRubToUsd(Long rub);
+    Long mapUsdToTenge(Long usd);
+    Long mapUsdToRub(Long usd);
 }
